created u32

+++ assembly/model.ts
@@ -10,12 +10,12 @@ export class PostedMessage {
   premium: boolean;
   sender: string;
   banned: boolean;
-  created: u64;
+  created: u32;
   constructor(public text: string) {
     assert(text.length <= maxLength, "Message is too long, max length is " + maxLength.toString());
     this.premium = context.attachedDeposit >= u128.from('10000000000000000000000');
     this.sender = context.sender;
-    this.created = (context.blockTimestamp - initDate) / 10 ** 9;
+    this.created = u32((context.blockTimestamp - initDate) / 10 ** 9);
   }
 }
 /**
